Снова переименовал методы, переделал EmployeeRowMApper - добавил проверку на 0l в getId, убрал двойные запросы в базу данных

+++ src/main/java/ru/supreme/webdemo/controller/EmployeeController.java
@@ -3,7 +3,9 @@ package ru.supreme.webdemo.controller;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import ru.supreme.webdemo.model.EmployeeEntity;
+import ru.supreme.webdemo.model.dto.EmployeeWithDepartmentIdDTO;
+import ru.supreme.webdemo.model.dto.EmployeeWithDepartmentNameDTO;
+import ru.supreme.webdemo.model.entity.EmployeeEntity;
 import ru.supreme.webdemo.service.EmployeeService;
 
 import java.util.List;
@@ -22,24 +24,53 @@ public class EmployeeController {
         this.employeeService = employeeService;
     }
 
-    @GetMapping(value = "/list")
-    public ResponseEntity<List<EmployeeEntity>> getAllEmployees() {
-        return ResponseEntity.status(HttpStatus.OK).body(employeeService.getAllEmployees());
+    @GetMapping
+    public ResponseEntity<List<EmployeeWithDepartmentNameDTO>> findAllEmployees(@RequestParam(name = "departmentId", required = false) Long id) {
+        if (id != null) {
+            List<EmployeeWithDepartmentNameDTO> list = employeeService.findEmployeesByDepartmentId(id);
+            if (list == null) {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+            } else {
+                return ResponseEntity.status(HttpStatus.OK).body(list);
+            }
+        } else {
+            return ResponseEntity.status(HttpStatus.OK).body(employeeService.findAllEmployees());
+        }
     }
 
-    @PostMapping(value = "/create")
-    public ResponseEntity<EmployeeEntity> createEmployee(@RequestBody EmployeeEntity employeeEntity) {
-        employeeService.save(employeeEntity);
-        return ResponseEntity.status(HttpStatus.CREATED).body(employeeEntity);
+    @GetMapping(value = "/{id}")
+    public ResponseEntity<EmployeeWithDepartmentNameDTO> findEmployeeById(@PathVariable(value = "id") Long id) {
+        EmployeeWithDepartmentNameDTO employeeWithDepartmentNameDTO = employeeService.findEmployeeById(id);
+        if (employeeWithDepartmentNameDTO == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        } else {
+            return ResponseEntity.status(HttpStatus.OK).body(employeeWithDepartmentNameDTO);
+        }
     }
+
+    @PostMapping
+    public ResponseEntity<EmployeeWithDepartmentIdDTO> create(@RequestBody EmployeeEntity employeeEntity) {
+        return ResponseEntity.status(HttpStatus.CREATED).body(employeeService.create(employeeEntity));
+    }
+
     @DeleteMapping(value = "/{id}")
-    public ResponseEntity<Void> deleteEmployee(@PathVariable(value = "id") Long id) {
-        employeeService.delete(id);
-        return ResponseEntity.status(HttpStatus.OK).build();
+    public ResponseEntity<Void> delete(@PathVariable(value = "id") Long id) {
+        if (employeeService.findEmployeeById(id) == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        } else {
+            employeeService.delete(id);
+            return ResponseEntity.status(HttpStatus.OK).build();
+        }
     }
-    @GetMapping
-    public ResponseEntity<EmployeeEntity> getEmployeeById(@RequestParam(value = "id") Long id) {
-        return ResponseEntity.status(HttpStatus.OK).body(employeeService.getEmployeeById(id));
+
+    @PatchMapping(value = "/{id}")
+    public ResponseEntity<EmployeeWithDepartmentIdDTO> update(@PathVariable(value = "id") Long id,
+                                                              @RequestBody EmployeeEntity employeeEntity) {
+        EmployeeWithDepartmentIdDTO employeeWithDepartmentIdDTO = employeeService.update(id, employeeEntity);
+        if (employeeWithDepartmentIdDTO == null) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        } else {
+            return ResponseEntity.status(HttpStatus.ACCEPTED).body(employeeWithDepartmentIdDTO);
+        }
     }
 }
-
